+++ expenses/admin.py
@@ -2,6 +2,8 @@ from django.contrib import admin
 
 from expenses.models import (
     Account,
+    Bank,
+    BankAccount,
     Currency,
     CurrencyConvert,
     Period,
@@ -27,6 +29,16 @@ class CurrencyConvertAdmin(admin.ModelAdmin):
 @admin.register(Account)
 class AccountAdmin(admin.ModelAdmin):
     ordering = ["name"]
+
+
+@admin.register(Bank)
+class BankAdmin(admin.ModelAdmin):
+    pass
+
+
+@admin.register(BankAccount)
+class BankAccountAdmin(admin.ModelAdmin):
+    pass
 
 
 @admin.register(Expense)